FamiTracker 0.4.6 initial pull.

+++ libs/famitracker/Source/common.h
@@ -22,12 +22,14 @@
 
 typedef unsigned char		uint8;
 typedef unsigned short		uint16;
-typedef unsigned int	    	uint32;
-typedef unsigned long    	uint64;
+typedef unsigned long		uint32;
+//typedef unsigned __int64	uint64;
+typedef unsigned long long	uint64;
 typedef signed char			int8;
 typedef signed short		int16;
-typedef signed int			int32;
-typedef signed long		int64;
+typedef signed long			int32;
+//typedef signed __int64		int64;
+typedef signed long long		int64;
 
 #define _MAIN_H_
